fix(web): stop the alerts strip asserting unmeasured channel health
deriveArmedState fed DND state and an active-alert count into a strip whose
documented contract is channel reachability. With no active alerts it printed
"All channels healthy. Alerts are armed." even if every channel was
unreachable, and one critical alert printed "Critical channels unreachable.
Alerts may not deliver." — a delivery-failure claim nothing had measured. The
function's own comment noted there is no per-channel health probe.

The strip becomes DndNoticeStrip: a single state carrying only what is known,
rendered only while a manual Do Not Disturb mute is in effect. Active alerts
and rule counts are already surfaced by the stat row and the active-alerts
card, so nothing is lost. "All channels healthy" also broke the no-verdicts
copy rule.

The strip's "Turn off" also cleared dndScheduleEnabled, silently deleting the
user's configured quiet hours; it now clears the manual mute only.
